requirements.txt
# =============================================================================
# OCX Python Services — Root Requirements
# Install all service dependencies in one shot:
#   pip install -r requirements.txt
#
# Or install per-service:
#   pip install -r trust-registry/requirements.txt
# =============================================================================

# --- Core Web Framework ---
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# --- gRPC ---
grpcio>=1.60.0
grpcio-tools>=1.60.0

# --- Databases ---
supabase>=2.3.0
psycopg2-binary>=2.9.9
redis>=5.0.0
elasticsearch>=8.11.0

# --- AI / LLM Clients ---
openai>=1.6.0
anthropic>=0.8.0

# --- Cryptography ---
ecdsa>=0.19.0

# --- Scientific Computing ---
numpy>=1.26.0
scipy>=1.11.0

# --- Reporting ---
reportlab>=4.0.0

# --- Policy Engine ---
json-logic-qubit>=0.9.1
pyyaml>=6.0.1

# --- Observability ---
prometheus-client>=0.19.0
